requirements.txt
tqdm
network
qrcode
bluedot
pyzbar
pillow